add validated-wrapped-abstract-control-superclass.ts
