+++ src/app/sections/payment-section/operations/invoices/invoices-list/invoice-details/create-payment-link-dialog/create-payment-link-dialog.component.ts
@@ -2,12 +2,12 @@ import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { FormControl } from '@ngneat/reactive-forms';
 import { BehaviorSubject, defer, merge, Subject } from 'rxjs';
-import { mapTo, shareReplay, switchMap } from 'rxjs/operators';
+import { mapTo, switchMap } from 'rxjs/operators';
 
 import { InvoiceService } from '@dsh/api';
 import { Controls } from '@dsh/app/shared/components/create-payment-link-form';
 import { CreatePaymentLinkService } from '@dsh/app/shared/services/create-payment-link/create-payment-link.service';
-import { SHARE_REPLAY_CONF } from '@dsh/operators';
+import { shareReplayRefCount } from '@dsh/operators';
 
 import { CreatePaymentLinkDialogData } from './types/create-payment-link-dialog-data';
 
@@ -17,9 +17,7 @@ import { CreatePaymentLinkDialogData } from './types/create-payment-link-dialog-
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CreatePaymentLinkDialogComponent {
-    paymentMethods$ = this.invoiceService
-        .getInvoicePaymentMethods(this.data.invoice.id)
-        .pipe(shareReplay(SHARE_REPLAY_CONF));
+    paymentMethods$ = this.invoiceService.getInvoicePaymentMethods(this.data.invoice.id).pipe(shareReplayRefCount());
 
     formControl = new FormControl<Controls>();
     paymentLink$ = merge(
@@ -31,8 +29,6 @@ export class CreatePaymentLinkDialogComponent {
         this.formControl.valueChanges.pipe(mapTo(''))
     );
     inProgress$ = new BehaviorSubject(false);
-    valid: boolean;
-    empty: boolean;
 
     private create$ = new Subject<void>();
