+++ src/utils/form/index.ts
@@ -1,5 +1,4 @@
 export * from './set-form-array-value';
 export * from './get-form-value-changes';
 export * from './get-form-validation-changes';
-export * from './wrapped-abstract-control-superclass';
-export * from './wrapped-form-group-superclass';
+export * from './validated-wrapped-abstract-control-superclass';
