+++ src/utils/form/validated-wrapped-abstract-control-superclass/create-validated-abstract-control-providers.ts
@@ -1,0 +1,10 @@
+import { ComponentType } from '@angular/cdk/overlay';
+import { Provider } from '@angular/core';
+import { provideValueAccessor } from '@s-libs/ng-core';
+
+import { provideValidator } from './provide-validator';
+
+export const createValidatedAbstractControlProviders = (component: ComponentType<unknown>): Provider[] => [
+    provideValueAccessor(component),
+    provideValidator(component),
+];
