+++ src/utils/form/validated-wrapped-abstract-control-superclass/index.ts
@@ -1,0 +1,4 @@
+export * from './wrapped-abstract-control-superclass';
+export * from './validated-wrapped-abstract-control-superclass';
+export { provideValidator } from './provide-validator';
+export { createValidatedAbstractControlProviders } from './create-validated-abstract-control-providers';
