+++ src/utils/form/validated-wrapped-abstract-control-superclass/provide-validator.ts
@@ -1,0 +1,9 @@
+import { ComponentType } from '@angular/cdk/overlay';
+import { Provider } from '@angular/core';
+import { NG_VALIDATORS } from '@angular/forms';
+
+export const provideValidator = (component: ComponentType<unknown>): Provider => ({
+    provide: NG_VALIDATORS,
+    useExisting: component,
+    multi: true,
+});
